Refactor worker flow.
