+++ index.html
@@ -17,7 +17,6 @@
 		<script src="src/util/clipper.js"></script>
 		<script src="src/util/parallel.js"></script>
 		<script src="src/util/geometryutil.js"></script>
-		<script src="src/util/placementworker.js"></script>
 		<script src="src/svgparser.js"></script>
 		<script src="src/genetic-algorithm/phenotype.js"></script>
 		<script src="src/genetic-algorithm/genetic-algorithm.js"></script>
