--- src/util/eval.js
@@ -1,4 +1,0 @@
-﻿this.onmessage = function (code) {
-  console.log(code);
-  eval(code.data);
-};
